Updated step 7. What to do if " ttyACM0 " belongs to the root user group
Updated step 7. What to do if " ttyACM0 " belongs to the root user group.
I noticed that in many Androids, "ttyACM0" belongs to the root user group, but we cannot manipulate the permissions of the root user group.
Based on this, I added step 7, which specifically handles the problem of "ttyACM0" belonging to the root user group.
This method may have a slight performance reduction compared to the processing method in step 6, but it ensures sufficient compatibility.